Net explorer search now in cytoscape explorer template.
Net explorer has been simplified: now instead of a search page and a results page,
there is only the result/explorer page. All the requests are now AJAX, and the elements are added
dynamically. This allows nodes from different DB to be displayed at the same time, for example, to be
able to compare different pathways in different transcriptomes/organisms.

Still to do:
	- "Node not found" response (now nothing happens).
	- Human nodes don't work yet.
	- Save graph to json.
	- Load json from file.
	- Tweak search form design.